refactor: co-locate staged directory parsing
